fix ldap userId with case problem (#2326)
fix ldap userId with case problem
